Convert ipc api files to typescript
