+++ src/renderer/window.ts
@@ -1,13 +1,10 @@
+import { ElectronAPI } from "@electron-toolkit/preload"
+import { API } from '../preload/index.ts'
+
 declare global {
   interface Window {
-    api: {
-      start: () => void,
-      loadState: () => string,
-      saveState: (state: string) => void,
-      downloadChapter: (mangaName: string, chapterNum: number, pages: string[], type: string) => void,
-      deleteChapter: (mangaName: string, chapterNum: number) => void,
-      deleteManga: (mangaName: string) => void,
-    }
+    electron: ElectronAPI,
+    api: API,
   }
 }
